make buffer data a reference type
mptcore
  buffer is responsible to track/terminate content lifetime
  make buffer instance more opaque
  require raw buffer for generic array data operations
  consisten tab ident in `array.h`
  add special buffer instances for typed array content

mpt++
  reorganize template array classes and content data

mptplot
  move typed_array to value elements

+++ mpt++/buffer.cpp
@@ -78,7 +78,7 @@ ssize_t Buffer::read(size_t nblk, void *dest, size_t esze)
 {
     Slice<uint8_t> d = data();
     if (!esze) {
-        if (d.length() < nblk) return -2;
+        if ((size_t) d.length() < nblk) return -2;
         if (dest) memcpy(dest, d.base(), nblk);
         return nblk;
     }
